#9553: create a separate enhnacer for handleLongText and use it for formatter table cell
