+++ web/client/utils/FeatureGridUtils.js
@@ -18,9 +18,7 @@ import {
 } from './ogc/WFS/base';
 
 import { applyDefaultToLocalizedString } from '../components/I18N/LocalizedString';
-import React from 'react';
-import OverlayTrigger from '../components/misc/OverlayTrigger';
-import { Tooltip } from 'react-bootstrap';
+import { handleLongTextEnhancer } from '../components/misc/enhancers/handleLongTextEnhancer';
 
 const getGeometryName = (describe) => get(findGeometryProperty(describe), "name");
 const getPropertyName = (name, describe) => name === "geometry" ? getGeometryName(describe) : name;
@@ -119,34 +117,6 @@ export const getCurrentPaginationOptions = ({ startPage, endPage }, oldPages, si
 };
 
 
-const formatterWrapperForLongContent = (props, RenderFormatter) => {
-    const { value } = props;
-    const cellRef = React.useRef(null);
-    const contentRef = React.useRef(null);
-    const [isContentOverflowing, setIsContentOverflowing] = React.useState(false);
-
-    const handleMouseEnter = () => {
-        const cellWidth = cellRef.current.offsetWidth;
-        const contentWidth = contentRef.current.offsetWidth;
-
-        if (contentWidth > cellWidth) {
-            setIsContentOverflowing(contentWidth > cellWidth);
-        } else setIsContentOverflowing(false);
-    };
-
-    return (<OverlayTrigger
-        placement="top"
-        overlay={isContentOverflowing ? <Tooltip id="tooltip">{RenderFormatter ? <RenderFormatter {...props} /> : value}</Tooltip> : <></>}
-    >
-        <div ref={cellRef} onMouseEnter={handleMouseEnter}>
-            <span ref={contentRef}>
-                <span>{RenderFormatter ? <RenderFormatter {...props} /> : value}</span>
-            </span>
-        </div>
-    </OverlayTrigger>);
-};
-
-
 /**
  * Utility function to get from a describeFeatureType response the columns to use in the react-data-grid
  * @param {object} describe describeFeatureType response
@@ -178,7 +148,7 @@ export const featureTypeToGridColumns = (
             editable,
             filterable,
             editor: getEditor(desc, field),
-            formatter: (props_) => formatterWrapperForLongContent(props_, getFormatter(desc, field)),
+            formatter: (props)=>handleLongTextEnhancer(props, getFormatter(desc, field)),
             filterRenderer: getFilterRenderer(desc, field)
         };
     });
